fix: use aigne hub model process env when run examples
