+++ models/aigne-hub/src/util/model.ts
@@ -11,12 +11,19 @@ import { nodejs } from "@aigne/platform-helpers/nodejs/index.js";
 import { XAIChatModel } from "@aigne/xai";
 import { NodeHttpHandler, streamCollector } from "@smithy/node-http-handler";
 import { HttpsProxyAgent } from "https-proxy-agent";
+import type inquirer from "inquirer";
 import type { ClientOptions } from "openai";
-import { joinURL } from "ufo";
-import { CliAIGNEHubChatModel } from "./cli-aigne-hub-model.js";
+import { AIGNEHubChatModel } from "../index.js";
+import {
+  AGENT_HUB_PROVIDER,
+  DEFAULT_AIGNE_HUB_PROVIDER_MODEL,
+  DEFAULT_MODEL_PROVIDER,
+  IsTest,
+} from "./constants.js";
+import { loadCredential } from "./credential.js";
+import type { LoadableModel, LoadCredentialOptions, Model } from "./type.js";
 
-const AIGNE_HUB_DID = "z8ia3xzq2tMq8CRHfaXj1BTYJyYnEcHbqP8cJ";
-export const AIGNE_HUB_URL = "https://hub.aigne.io/";
+const { MODEL_PROVIDER, MODEL_NAME } = nodejs.env;
 
 export function availableModels(): LoadableModel[] {
   const proxy = ["HTTPS_PROXY", "https_proxy", "HTTP_PROXY", "http_proxy", "ALL_PROXY", "all_proxy"]
@@ -80,22 +87,11 @@ export function availableModels(): LoadableModel[] {
       create: (params) => new XAIChatModel({ ...params, clientOptions }),
     },
     {
-      name: CliAIGNEHubChatModel.name,
+      name: AIGNEHubChatModel.name,
       apiKeyEnvName: "AIGNE_HUB_API_KEY",
-      create: (params) => new CliAIGNEHubChatModel({ ...params, clientOptions }),
+      create: (params) => new AIGNEHubChatModel({ ...params, clientOptions }),
     },
   ];
-}
-
-export interface LoadableModel {
-  name: string | string[];
-  apiKeyEnvName?: string | string[];
-  create: (options: {
-    model?: string;
-    modelOptions?: ChatModelOptions;
-    apiKey?: string;
-    url?: string;
-  }) => ChatModel;
 }
 
 export function findModel(models: LoadableModel[], provider: string) {
@@ -107,24 +103,66 @@ export function findModel(models: LoadableModel[], provider: string) {
   });
 }
 
-const { MODEL_PROVIDER, MODEL_NAME } = nodejs.env;
-const DEFAULT_MODEL_PROVIDER = "openai";
+export const parseModelOption = (model?: string) => {
+  const { provider, name } =
+    (model || process.env.MODEL)?.match(/(?<provider>[^:]*)(:(?<name>(\S+)))?/)?.groups ?? {};
 
-type Model =
-  | {
-      provider?: string | null;
-      name?: string | null;
-      temperature?: number | null;
-      topP?: number | null;
-      frequencyPenalty?: number | null;
-      presencePenalty?: number | null;
-    }
-  | undefined;
+  return { provider, name };
+};
+
+export const formatModelName = async (model: string, inquirerPrompt: typeof inquirer.prompt) => {
+  const models = availableModels();
+  if (!model) return DEFAULT_AIGNE_HUB_PROVIDER_MODEL;
+
+  const { provider, name } = parseModelOption(model);
+
+  if (!provider) {
+    return DEFAULT_AIGNE_HUB_PROVIDER_MODEL;
+  }
+
+  const providerName = provider.replace(/-/g, "");
+  if (providerName.includes(AGENT_HUB_PROVIDER)) {
+    return model;
+  }
+
+  const m = findModel(models, providerName);
+  if (!m) throw new Error(`Unsupported model: ${provider} ${name}`);
+
+  const apiKeyEnvName = Array.isArray(m.apiKeyEnvName) ? m.apiKeyEnvName : [m.apiKeyEnvName];
+  if (apiKeyEnvName.some((name) => name && process.env[name])) {
+    return model;
+  }
+
+  if (IsTest) {
+    return `${AGENT_HUB_PROVIDER}:${provider}/${name}`;
+  }
+
+  const result = await inquirerPrompt({
+    type: "list",
+    name: "useAigneHub",
+    message: `Seems no API Key configured for ${provider}/${name}, select your preferred way to continue:`,
+    choices: [
+      {
+        name: `Connect to AIGNE Hub to use ${name} (Recommended since free credits available)`,
+        value: true,
+      },
+      {
+        name: `Exit and bring my owner API Key by set ${apiKeyEnvName.join(", ")}`,
+        value: false,
+      },
+    ],
+    default: true,
+  });
+
+  if (!result.useAigneHub) return model;
+
+  return `${AGENT_HUB_PROVIDER}:${provider}/${name}`;
+};
 
 export async function loadModel(
   model?: Model,
   modelOptions?: ChatModelOptions,
-  credential?: { apiKey?: string; url?: string },
+  options?: LoadCredentialOptions,
 ): Promise<ChatModel | undefined> {
   const params = {
     model: MODEL_NAME ?? model?.name ?? undefined,
@@ -135,25 +173,16 @@ export async function loadModel(
   };
 
   const provider = (MODEL_PROVIDER ?? model?.provider ?? DEFAULT_MODEL_PROVIDER).replace(/-/g, "");
+  const models = availableModels();
 
-  const m = findModel(availableModels(), provider);
+  const m = findModel(models, provider);
   if (!m) throw new Error(`Unsupported model: ${model?.provider} ${model?.name}`);
+
+  const credential = await loadCredential({ ...options, model: `${provider}:${params.model}` });
 
   return m.create({
     ...(credential || {}),
     model: params.model,
     modelOptions: { ...params, ...modelOptions },
   });
-}
-
-export async function getAIGNEHubMountPoint(url: string) {
-  const { origin } = new URL(url);
-  const BLOCKLET_JSON_PATH = "__blocklet__.js?type=json";
-  const blockletInfo = await fetch(joinURL(origin, BLOCKLET_JSON_PATH));
-  const blocklet = await blockletInfo.json();
-  const aigneHubMount = (blocklet?.componentMountPoints || []).find(
-    (m: { did: string }) => m.did === AIGNE_HUB_DID,
-  );
-
-  return joinURL(origin, aigneHubMount?.mountPoint || "");
 }
