+++ packages/cli/package.json
@@ -70,10 +70,8 @@
     "@inquirer/type": "^3.0.8",
     "@listr2/prompt-adapter-inquirer": "^3.0.1",
     "@modelcontextprotocol/sdk": "^1.15.0",
-    "@ocap/mcrypto": "^1.21.0",
     "@smithy/node-http-handler": "^4.1.0",
     "chalk": "^5.4.1",
-    "crypto": "^1.0.1",
     "detect-port": "^2.1.0",
     "dotenv-flow": "^4.1.0",
     "express": "^5.1.0",
@@ -84,9 +82,7 @@
     "log-update": "^6.1.0",
     "marked": "^16.0.0",
     "nunjucks": "^3.2.4",
-    "open": "10.1.2",
     "openai": "^5.8.3",
-    "p-wait-for": "^5.0.2",
     "prettier": "^3.6.2",
     "tar": "^7.4.3",
     "wrap-ansi": "^9.0.0",
