+++ packages/cli/src/utils/load-aigne.ts
@@ -1,51 +1,24 @@
 import { existsSync, mkdirSync } from "node:fs";
-import { readFile, writeFile } from "node:fs/promises";
+import { readFile } from "node:fs/promises";
 import { homedir } from "node:os";
 import { join } from "node:path";
-import { availableModels, findModel, loadModel } from "@aigne/aigne-hub";
-import { AIGNE } from "@aigne/core";
+import type { LoadCredentialOptions, Model } from "@aigne/aigne-hub";
+import {
+  AIGNE_ENV_FILE,
+  checkConnectionStatus,
+  AIGNE_HUB_URL as DEFAULT_AIGNE_HUB_URL,
+  formatModelName,
+  loadModel,
+  parseModelOption,
+} from "@aigne/aigne-hub";
+import { AIGNE, type ChatModelOptions } from "@aigne/core";
 import { loadAIGNEFile } from "@aigne/core/loader/index.js";
-import { logger } from "@aigne/core/utils/logger.js";
-import { AesCrypter } from "@ocap/mcrypto/lib/crypter/aes-legacy.js";
-import crypto from "crypto";
 import inquirer from "inquirer";
-import open from "open";
-import pWaitFor from "p-wait-for";
-import { joinURL, withQuery } from "ufo";
 import { parse, stringify } from "yaml";
 import { availableMemories } from "../constants.js";
-import { parseModelOption, type RunAIGNECommandOptions } from "./run-with-aigne.js";
+import type { RunAIGNECommandOptions } from "./run-with-aigne.js";
 
-const aes = new AesCrypter();
-export const decrypt = (m: string, s: string, i: string) =>
-  aes.decrypt(m, crypto.pbkdf2Sync(i, s, 256, 32, "sha512").toString("hex"));
-export const encrypt = (m: string, s: string, i: string) =>
-  aes.encrypt(m, crypto.pbkdf2Sync(i, s, 256, 32, "sha512").toString("hex"));
-
-const escapeFn = (str: string) => str.replace(/\+/g, "-").replace(/\//g, "_").replace(/=/g, "");
-const unescapeFn = (str: string) =>
-  (str + "===".slice((str.length + 3) % 4)).replace(/-/g, "+").replace(/_/g, "/");
-export const encodeEncryptionKey = (key: string) => escapeFn(Buffer.from(key).toString("base64"));
-export const decodeEncryptionKey = (str: string) =>
-  new Uint8Array(Buffer.from(unescapeFn(str), "base64"));
-
-const TEST_ENV = process.env.CI || process.env.NODE_ENV === "test";
-export const AIGNE_ENV_FILE = TEST_ENV
-  ? join(homedir(), ".aigne", "test-aigne-hub-connected.yaml")
-  : join(homedir(), ".aigne", "aigne-hub-connected.yaml");
-
-const request = async (config: { url: string; method?: string; requestCount?: number }) => {
-  const headers: Record<string, string> = {};
-  if (config.requestCount !== undefined) {
-    headers["X-Request-Count"] = config.requestCount.toString();
-  }
-
-  const response = await fetch(config.url, { method: config.method || "GET", headers });
-  if (!response.ok) throw new Error(`HTTP error! status: ${response.status}`);
-
-  const data = await response.json();
-  return { data };
-};
+const IsTest = process.env.CI || process.env.NODE_ENV === "test";
 
 export interface RunOptions extends RunAIGNECommandOptions {
   path: string;
@@ -54,271 +27,13 @@ export interface RunOptions extends RunAIGNECommandOptions {
   aigneHubUrl?: string;
 }
 
-const WELLKNOWN_SERVICE_PATH_PREFIX = "/.well-known/service";
-const ACCESS_KEY_PREFIX = "/api/access-key";
-const ACCESS_KEY_SESSION_API = `${ACCESS_KEY_PREFIX}/session`;
-
-type FetchResult = { accessKeyId: string; accessKeySecret: string };
-
-export const fetchConfigs = async ({
-  connectUrl,
-  sessionId,
-  fetchInterval,
-  fetchTimeout,
-}: {
-  connectUrl: string;
-  sessionId: string;
-  fetchInterval: number;
-  fetchTimeout: number;
-}) => {
-  const sessionURL = withQuery(joinURL(connectUrl, ACCESS_KEY_SESSION_API), { sid: sessionId });
-
-  let requestCount = 0;
-  const condition = async () => {
-    const { data: session } = await request({ url: sessionURL, requestCount });
-    requestCount++;
-    return Boolean(session.accessKeyId && session.accessKeySecret);
-  };
-
-  await pWaitFor(condition, { interval: fetchInterval, timeout: fetchTimeout });
-
-  const { data: session } = await request({ url: sessionURL, requestCount });
-  await request({ url: sessionURL, method: "DELETE" });
-
-  return {
-    ...session,
-    accessKeyId: session.accessKeyId,
-    accessKeySecret: decrypt(session.accessKeySecret, session.accessKeyId, session.challenge),
-  };
-};
-
-function baseWrapSpinner(_: string, waiting: () => Promise<FetchResult>) {
-  return Promise.resolve(waiting());
-}
-
-interface CreateConnectOptions {
-  connectUrl: string;
-  openPage?: (url: string) => void;
-  fetchInterval?: number;
-  retry?: number;
-  source?: string;
-  connectAction?: string;
-  appName?: string;
-  appLogo?: string;
-  wrapSpinner?: typeof baseWrapSpinner;
-  prettyUrl?: (url: string) => string;
-  closeOnSuccess?: boolean;
-  intervalFetchConfig?: (options: {
-    sessionId: string;
-    fetchInterval: number;
-    fetchTimeout: number;
-  }) => Promise<FetchResult>;
-}
-
-export async function createConnect({
-  connectUrl,
-  openPage,
-  fetchInterval = 3 * 1000,
-  retry = 1500,
-  source = "Blocklet CLI",
-  connectAction = "connect-cli",
-  wrapSpinner = baseWrapSpinner,
-  closeOnSuccess,
-  intervalFetchConfig,
-  appName = "AIGNE CLI",
-  appLogo = "https://www.aigne.io/favicon.ico?imageFilter=resize&w=32",
-}: CreateConnectOptions) {
-  try {
-    const startSessionURL = joinURL(connectUrl, ACCESS_KEY_SESSION_API);
-    const { data: session } = await request({ url: startSessionURL, method: "POST" });
-    const token = session.id;
-
-    const pageUrl = withQuery(joinURL(connectUrl, connectAction), {
-      __token__: encodeEncryptionKey(token),
-      source,
-      closeOnSuccess,
-      cli: true,
-      appName: ` ${appName}`,
-      appLogo,
-    });
-
-    openPage?.(pageUrl);
-
-    return await wrapSpinner(`Waiting for connection: ${connectUrl}`, async () => {
-      const checkAuthorizeStatus = intervalFetchConfig ?? fetchConfigs;
-
-      const authorizeStatus = await checkAuthorizeStatus({
-        connectUrl,
-        sessionId: token,
-        fetchTimeout: retry * fetchInterval,
-        fetchInterval: retry,
-      });
-
-      return authorizeStatus;
-    });
-  } catch (e) {
-    console.error(e);
-    throw e;
-  }
-}
-
-const AGENT_HUB_PROVIDER = "aignehub";
-const DEFAULT_AIGNE_HUB_MODEL = "openai/gpt-4o";
-const DEFAULT_AIGNE_HUB_PROVIDER_MODEL = `${AGENT_HUB_PROVIDER}:${DEFAULT_AIGNE_HUB_MODEL}`;
-export const DEFAULT_URL = "https://hub.aigne.io/";
-
-export const formatModelName = async (
-  models: ReturnType<typeof availableModels>,
-  model: string,
-  inquirerPrompt: typeof inquirer.prompt,
-) => {
-  if (!model) return DEFAULT_AIGNE_HUB_PROVIDER_MODEL;
-
-  const { provider, name } = parseModelOption(model);
-
-  if (!provider) {
-    return DEFAULT_AIGNE_HUB_PROVIDER_MODEL;
-  }
-
-  const providerName = provider.replace(/-/g, "");
-  if (providerName.includes(AGENT_HUB_PROVIDER)) {
-    return model;
-  }
-
-  const m = findModel(models, providerName);
-  if (!m) throw new Error(`Unsupported model: ${provider} ${name}`);
-
-  const apiKeyEnvName = Array.isArray(m.apiKeyEnvName) ? m.apiKeyEnvName : [m.apiKeyEnvName];
-  if (apiKeyEnvName.some((name) => name && process.env[name])) {
-    return model;
-  }
-
-  if (TEST_ENV) {
-    return `${AGENT_HUB_PROVIDER}:${provider}/${name}`;
-  }
-
-  const result = await inquirerPrompt({
-    type: "list",
-    name: "useAigneHub",
-    message: `Seems no API Key configured for ${provider}/${name}, select your preferred way to continue:`,
-    choices: [
-      {
-        name: `Connect to AIGNE Hub to use ${name} (Recommended since free credits available)`,
-        value: true,
-      },
-      {
-        name: `Exit and bring my owner API Key by set ${apiKeyEnvName.join(", ")}`,
-        value: false,
-      },
-    ],
-    default: true,
-  });
-
-  if (!result.useAigneHub) return model;
-
-  return `${AGENT_HUB_PROVIDER}:${provider}/${name}`;
-};
-
-export async function connectToAIGNEHub(url: string) {
-  const { origin, host } = new URL(url);
-  const connectUrl = joinURL(origin, WELLKNOWN_SERVICE_PATH_PREFIX);
-
-  const BLOCKLET_JSON_PATH = "__blocklet__.js?type=json";
-  const blockletInfo = await fetch(joinURL(origin, BLOCKLET_JSON_PATH));
-  const blocklet = await blockletInfo.json();
-  const aigneHubMount = (blocklet?.componentMountPoints || []).find(
-    (m: { did: string }) => m.did === "z8ia3xzq2tMq8CRHfaXj1BTYJyYnEcHbqP8cJ",
-  );
-
-  try {
-    const result = await createConnect({
-      connectUrl: connectUrl,
-      connectAction: "gen-simple-access-key",
-      source: `@aigne/cli connect to AIGNE hub`,
-      closeOnSuccess: true,
-      openPage: (pageUrl) => open(pageUrl),
-    });
-
-    const accessKeyOptions = {
-      apiKey: result.accessKeySecret,
-      url: joinURL(origin, aigneHubMount?.mountPoint || ""),
-    };
-
-    // After redirection, write the AIGNE Hub access token
-    const aigneDir = join(homedir(), ".aigne");
-    if (!existsSync(aigneDir)) {
-      mkdirSync(aigneDir, { recursive: true });
-    }
-
-    const envs = parse(await readFile(AIGNE_ENV_FILE, "utf8").catch(() => stringify({})));
-
-    await writeFile(
-      AIGNE_ENV_FILE,
-      stringify({
-        ...envs,
-        [host]: {
-          AIGNE_HUB_API_KEY: accessKeyOptions.apiKey,
-          AIGNE_HUB_API_URL: accessKeyOptions.url,
-        },
-        default: {
-          AIGNE_HUB_API_URL: accessKeyOptions.url,
-        },
-      }),
-    ).catch((err) => {
-      logger.error(
-        "Failed to write AIGNE Hub access token to .aigne/aigne-hub-connected.yaml",
-        err.message,
-      );
-      throw err;
-    });
-
-    return accessKeyOptions;
-  } catch (error) {
-    logger.error("Failed to connect to AIGNE Hub", error.message);
-    return { apiKey: undefined, url: undefined };
-  }
-}
-
-export const checkConnectionStatus = async (host: string) => {
-  // aigne-hub access token
-  if (!existsSync(AIGNE_ENV_FILE)) {
-    throw new Error("AIGNE_HUB_API_KEY file not found, need to login first");
-  }
-
-  const data = await readFile(AIGNE_ENV_FILE, "utf8");
-  if (!data.includes("AIGNE_HUB_API_KEY")) {
-    throw new Error("AIGNE_HUB_API_KEY key not found, need to login first");
-  }
-
-  const envs = parse(data);
-  if (!envs[host]) {
-    throw new Error("AIGNE_HUB_API_KEY host not found, need to login first");
-  }
-
-  const env = envs[host];
-  if (!env.AIGNE_HUB_API_KEY) {
-    throw new Error("AIGNE_HUB_API_KEY key not found, need to login first");
-  }
-
-  return {
-    apiKey: env.AIGNE_HUB_API_KEY,
-    url: joinURL(env.AIGNE_HUB_API_URL),
-  };
-};
-
 const mockInquirerPrompt = (() => Promise.resolve({ useAigneHub: true })) as any;
 
-export async function loadAIGNE(
-  path: string,
-  options?: Pick<RunOptions, "model" | "aigneHubUrl">,
+export async function loadAIGNEByOptions(
+  options?: Model & Pick<RunOptions, "model" | "aigneHubUrl">,
+  modelOptions?: ChatModelOptions,
   actionOptions?: {
-    inquirerPromptFn?: (prompt: {
-      type: string;
-      name: string;
-      message: string;
-      choices: { name: string; value: any }[];
-      default: any;
-    }) => Promise<any>;
+    inquirerPromptFn?: LoadCredentialOptions["inquirerPromptFn"];
     runTest?: boolean;
   },
 ) {
@@ -331,90 +46,77 @@ export async function loadAIGNE(
   const inquirerPrompt = (actionOptions?.inquirerPromptFn ??
     inquirer.prompt) as typeof inquirer.prompt;
 
-  const models = availableModels();
+  // get aigne hub url from options or env
   const configUrl = options?.aigneHubUrl || process.env.AIGNE_HUB_API_URL;
-  const AIGNE_HUB_URL = configUrl || envs?.default?.AIGNE_HUB_API_URL || DEFAULT_URL;
+  const AIGNE_HUB_URL = configUrl || envs?.default?.AIGNE_HUB_API_URL || DEFAULT_AIGNE_HUB_URL;
 
-  const connectUrl = joinURL(new URL(AIGNE_HUB_URL).origin, WELLKNOWN_SERVICE_PATH_PREFIX);
+  const { host } = new URL(AIGNE_HUB_URL);
+
+  const result = await checkConnectionStatus(host).catch(() => null);
+  const alreadyConnected = Boolean(result?.apiKey);
+
+  // format model name
+  const modelName = await formatModelName(
+    options?.model || "",
+    alreadyConnected ? mockInquirerPrompt : inquirerPrompt,
+  );
+
+  const model = await loadModel(
+    {
+      ...parseModelOption(modelName),
+      temperature: options?.temperature,
+      topP: options?.topP,
+      presencePenalty: options?.presencePenalty,
+      frequencyPenalty: options?.frequencyPenalty,
+    },
+    modelOptions,
+    { aigneHubUrl: AIGNE_HUB_URL, inquirerPromptFn: actionOptions?.inquirerPromptFn },
+  );
+
+  return new AIGNE({ model });
+}
+
+export async function loadAIGNE(
+  path: string,
+  options?: Pick<RunOptions, "model" | "aigneHubUrl">,
+  actionOptions?: {
+    inquirerPromptFn?: LoadCredentialOptions["inquirerPromptFn"];
+    runTest?: boolean;
+  },
+) {
+  const aigneDir = join(homedir(), ".aigne");
+  if (!existsSync(aigneDir)) {
+    mkdirSync(aigneDir, { recursive: true });
+  }
+
+  const envs = parse(await readFile(AIGNE_ENV_FILE, "utf8").catch(() => stringify({})));
+  const inquirerPrompt = (actionOptions?.inquirerPromptFn ??
+    inquirer.prompt) as typeof inquirer.prompt;
+
+  // get aigne hub url from options or env
+  const configUrl = options?.aigneHubUrl || process.env.AIGNE_HUB_API_URL;
+  const AIGNE_HUB_URL = configUrl || envs?.default?.AIGNE_HUB_API_URL || DEFAULT_AIGNE_HUB_URL;
 
   const { host } = new URL(AIGNE_HUB_URL);
   const { aigne } = await loadAIGNEFile(path).catch(() => ({ aigne: null }));
 
   const result = await checkConnectionStatus(host).catch(() => null);
   const alreadyConnected = Boolean(result?.apiKey);
+
+  // format model name
   const modelName = await formatModelName(
-    models,
     options?.model || `${aigne?.model?.provider ?? ""}:${aigne?.model?.name ?? ""}`,
     alreadyConnected ? mockInquirerPrompt : inquirerPrompt,
   );
 
-  let credential: { apiKey?: string; url?: string } = {};
-
-  if (TEST_ENV && !actionOptions?.runTest) {
+  if (IsTest && !actionOptions?.runTest) {
     const model = await loadModel(parseModelOption(modelName));
     return await AIGNE.load(path, { loadModel, memories: availableMemories, model });
   }
 
-  if ((modelName.toLocaleLowerCase() || "").includes(AGENT_HUB_PROVIDER)) {
-    try {
-      credential = await checkConnectionStatus(host);
-    } catch (error) {
-      if (error instanceof Error && error.message.includes("login first")) {
-        let aigneHubUrl = connectUrl;
-
-        if (!configUrl) {
-          const { subscribe } = await inquirerPrompt({
-            type: "list",
-            name: "subscribe",
-            message:
-              "No LLM API Keys or AIGNE Hub connections found. How would you like to proceed?",
-            choices: [
-              {
-                name: "Connect to the Arcblock official AIGNE Hub (recommended, free credits for new users)",
-                value: "official",
-              },
-              connectUrl.includes(DEFAULT_URL)
-                ? {
-                    name: "Connect to your own AIGNE Hub instance (self-hosted)",
-                    value: "custom",
-                  }
-                : null,
-              {
-                name: "Exit and configure my own LLM API Keys",
-                value: "manual",
-              },
-            ].filter(Boolean) as { name: string; value: string }[],
-            default: "official",
-          });
-
-          if (subscribe === "custom") {
-            const { customUrl } = await inquirerPrompt({
-              type: "input",
-              name: "customUrl",
-              message: "Enter the URL of your AIGNE Hub:",
-              validate(input) {
-                try {
-                  const url = new URL(input);
-                  return url.protocol.startsWith("http")
-                    ? true
-                    : "Must be a valid URL with http or https";
-                } catch {
-                  return "Invalid URL";
-                }
-              },
-            });
-            aigneHubUrl = customUrl;
-          } else if (subscribe === "manual") {
-            console.log("You chose to configure your own LLM API Keys. Exiting...");
-            process.exit(0);
-          }
-        }
-
-        credential = await connectToAIGNEHub(aigneHubUrl);
-      }
-    }
-  }
-
-  const model = await loadModel(parseModelOption(modelName), undefined, credential);
+  const model = await loadModel(parseModelOption(modelName), undefined, {
+    aigneHubUrl: AIGNE_HUB_URL,
+    inquirerPromptFn: actionOptions?.inquirerPromptFn,
+  });
   return await AIGNE.load(path, { loadModel, memories: availableMemories, model });
 }
